Move test_starting.py into tuning/
