+++ pymc/tests/tuning/test_starting.py
@@ -14,9 +14,15 @@
 import numpy as np
 import pytest
 
+import pymc as pm
+
+from pymc.exceptions import ImputationWarning
 from pymc.step_methods.metropolis import tune
 from pymc.tests import models
-from pymc.tuning import find_MAP
+from pymc.tests.checks import close_to
+from pymc.tests.helpers import select_by_precision
+from pymc.tests.models import non_normal, simple_arbitrary_det, simple_model
+from pymc.tuning import find_MAP, starting
 
 
 @pytest.mark.parametrize("bounded", [False, True])
@@ -36,3 +42,133 @@ def test_tune_not_inplace():
     returned_scaling = tune(orig_scaling, acc_rate=0.6)
     assert not returned_scaling is orig_scaling
     assert np.all(orig_scaling == np.array([0.001, 0.1]))
+
+
+def test_accuracy_normal():
+    _, model, (mu, _) = simple_model()
+    with model:
+        newstart = find_MAP(pm.Point(x=[-10.5, 100.5]))
+        close_to(newstart["x"], [mu, mu], select_by_precision(float64=1e-5, float32=1e-4))
+
+
+def test_accuracy_non_normal():
+    _, model, (mu, _) = non_normal(4)
+    with model:
+        newstart = find_MAP(pm.Point(x=[0.5, 0.01, 0.95, 0.99]))
+        close_to(newstart["x"], mu, select_by_precision(float64=1e-5, float32=1e-4))
+
+
+def test_find_MAP_discrete():
+    tol1 = 2.0**-11
+    tol2 = 2.0**-6
+    alpha = 4
+    beta = 4
+    n = 20
+    yes = 15
+
+    with pm.Model() as model:
+        p = pm.Beta("p", alpha, beta)
+        pm.Binomial("ss", n=n, p=p)
+        pm.Binomial("s", n=n, p=p, observed=yes)
+
+        map_est1 = find_MAP()
+        map_est2 = find_MAP(vars=model.value_vars)
+
+    close_to(map_est1["p"], 0.6086956533498806, tol1)
+
+    close_to(map_est2["p"], 0.695642178810167, tol2)
+    assert map_est2["ss"] == 14
+
+
+def test_find_MAP_no_gradient():
+    _, model = simple_arbitrary_det()
+    with model:
+        find_MAP()
+
+
+def test_find_MAP():
+    tol = 2.0**-11  # 16 bit machine epsilon, a low bar
+    data = np.random.randn(100)
+    # data should be roughly mean 0, std 1, but let's
+    # normalize anyway to get it really close
+    data = (data - np.mean(data)) / np.std(data)
+
+    with pm.Model():
+        mu = pm.Uniform("mu", -1, 1)
+        sigma = pm.Uniform("sigma", 0.5, 1.5)
+        pm.Normal("y", mu=mu, tau=sigma**-2, observed=data)
+
+        # Test gradient minimization
+        map_est1 = find_MAP(progressbar=False)
+        # Test non-gradient minimization
+        map_est2 = find_MAP(progressbar=False, method="Powell")
+
+    close_to(map_est1["mu"], 0, tol)
+    close_to(map_est1["sigma"], 1, tol)
+
+    close_to(map_est2["mu"], 0, tol)
+    close_to(map_est2["sigma"], 1, tol)
+
+
+def test_find_MAP_issue_5923():
+    # Test that gradient-based minimization works well regardless of the order
+    # of variables in `vars`, and even when starting a reasonable distance from
+    # the MAP.
+    tol = 2.0**-11  # 16 bit machine epsilon, a low bar
+    data = np.random.randn(100)
+    # data should be roughly mean 0, std 1, but let's
+    # normalize anyway to get it really close
+    data = (data - np.mean(data)) / np.std(data)
+
+    with pm.Model():
+        mu = pm.Uniform("mu", -1, 1)
+        sigma = pm.Uniform("sigma", 0.5, 1.5)
+        pm.Normal("y", mu=mu, tau=sigma**-2, observed=data)
+
+        start = {"mu": -0.5, "sigma": 1.25}
+        map_est1 = find_MAP(progressbar=False, vars=[mu, sigma], start=start)
+        map_est2 = find_MAP(progressbar=False, vars=[sigma, mu], start=start)
+
+    close_to(map_est1["mu"], 0, tol)
+    close_to(map_est1["sigma"], 1, tol)
+
+    close_to(map_est2["mu"], 0, tol)
+    close_to(map_est2["sigma"], 1, tol)
+
+
+def test_find_MAP_issue_4488():
+    # Test for https://github.com/pymc-devs/pymc/issues/4488
+    with pm.Model() as m:
+        with pytest.warns(ImputationWarning):
+            x = pm.Gamma("x", alpha=3, beta=10, observed=np.array([1, np.nan]))
+        y = pm.Deterministic("y", x + 1)
+        map_estimate = find_MAP()
+
+    assert not set.difference({"x_missing", "x_missing_log__", "y"}, set(map_estimate.keys()))
+    np.testing.assert_allclose(map_estimate["x_missing"], 0.2, rtol=1e-4, atol=1e-4)
+    np.testing.assert_allclose(map_estimate["y"], [2.0, map_estimate["x_missing"][0] + 1])
+
+
+def test_allinmodel():
+    model1 = pm.Model()
+    model2 = pm.Model()
+    with model1:
+        x1 = pm.Normal("x1", mu=0, sigma=1)
+        y1 = pm.Normal("y1", mu=0, sigma=1)
+    with model2:
+        x2 = pm.Normal("x2", mu=0, sigma=1)
+        y2 = pm.Normal("y2", mu=0, sigma=1)
+
+    x1 = model1.rvs_to_values[x1]
+    y1 = model1.rvs_to_values[y1]
+    x2 = model2.rvs_to_values[x2]
+    y2 = model2.rvs_to_values[y2]
+
+    starting.allinmodel([x1, y1], model1)
+    starting.allinmodel([x1], model1)
+    with pytest.raises(ValueError, match=r"Some variables not in the model: \['x2', 'y2'\]"):
+        starting.allinmodel([x2, y2], model1)
+    with pytest.raises(ValueError, match=r"Some variables not in the model: \['x2'\]"):
+        starting.allinmodel([x2, y1], model1)
+    with pytest.raises(ValueError, match=r"Some variables not in the model: \['x2'\]"):
+        starting.allinmodel([x2], model1)
